01.Reto01/Nodo4/requirements.txt
blinker
certifi
charset-normalizer
click
Flask
grpcio
grpcio-tools
idna
itsdangerous
Jinja2
MarkupSafe
pip
protobuf==5.27.2
requests
setuptools
urllib3
Werkzeug
ujson
pandas